+++ packages/core/CHANGELOG.md
@@ -4,6 +4,19 @@
 
 ### Minor Changes
 
+- Bump all main packages to version 0.3.3 for release consistency.
+
+### Patch Changes
+
+- Updated dependencies
+  - @aevatar-react-sdk/utils@0.4.0
+  - @aevatar-react-sdk/types@0.4.0
+  - @aevatar-react-sdk/services@0.4.0
+
+## 0.3.1
+
+### Minor Changes
+
 - 729ba97: change input to textaere
 - 37b6a1e: update
 
@@ -11,9 +24,9 @@
 
 - Updated dependencies [729ba97]
 - Updated dependencies [37b6a1e]
-  - @aevatar-react-sdk/services@0.4.0
-  - @aevatar-react-sdk/types@0.4.0
-  - @aevatar-react-sdk/utils@0.4.0
+  - @aevatar-react-sdk/services@0.3.1
+  - @aevatar-react-sdk/types@0.3.1
+  - @aevatar-react-sdk/utils@0.3.1
 
 ## 0.3.1-alpha.0