+++ packages/utils/CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ### Minor Changes
 
+- Bump all main packages to version 0.3.3 for release consistency.
+
+### Patch Changes
+
+- Updated dependencies
+  - @aevatar-react-sdk/types@0.4.0
+
+## 0.3.1
+
+### Minor Changes
+
 - 729ba97: change input to textaere
 - 37b6a1e: update